% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/idin.R
\name{\%idin\%}
\alias{\%idin\%}
\title{this is R's original %in% function}
\usage{
x \%idin\% y
}
\arguments{
\item{x:}{vector for which you're looking for matches}

\item{y:}{vector you're looking for matches in}
}
\value{
A logical vector of length x
}
\description{
this is R's original %in% function
}
\examples{
x <- seq(from = 1, to = 3, by = 0.1)
y <- c(3, 2.4, 1.8, 1.4)
x \%idin\% y
}
