% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/in.R
\name{\%in\%}
\alias{\%in\%}
\title{replacing the %in% function with the all.equal equivalent}
\usage{
x \%in\% y
}
\arguments{
\item{x:}{vector for which you're looking for matches}

\item{y:}{vector you're looking for matches in}
}
\value{
A logical vector
}
\description{
replacing the %in% function with the all.equal equivalent
}
\examples{
x <- seq(from = 1, to = 3, by = 0.1)
y <- c(3, 2.4, 1.8, 1.4)
x \%in\% y
}
